+++ examples/browser-repl/src/shared.cljs
@@ -1,0 +1,6 @@
+(ns shared)
+
+(defn shared-function []
+  (+ 1 2 3))
+
+(defmulti shape-label :type)