+++ examples/replicant/vite.config.js
@@ -1,0 +1,9 @@
+// vite.config.js
+import { defineConfig } from 'vite';
+import cherry from 'cherry-cljs/vite.js';
+
+export default defineConfig(() => {
+  return {
+    plugins: [cherry()],
+  };
+});